Feature/typescript for life

+++ client/src/features/goals/goalSlice.tsx
@@ -1,7 +1,20 @@
-import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { RootState } from "../../app/store";
 import goalService from "./goalService";
 
-const initialState = {
+export type Goal = {
+  goals: {
+    _id: string;
+    text: string;
+    createdAt: string;
+  }[];
+  isError: boolean;
+  isSuccess: boolean;
+  isLoading: boolean;
+  message: unknown;
+};
+
+const initialState: Goal = {
   goals: [],
   isError: false,
   isSuccess: false,
@@ -14,8 +27,8 @@ export const createGoal = createAsyncThunk(
   "goals/create",
   async (goalData, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.createGoal(goalData, token);
+      const token = (thunkAPI.getState() as RootState)?.auth?.user?.token;
+      if (token) return await goalService.createGoal(goalData, token);
     } catch (error: any) {
       const message =
         (error.response &&
@@ -34,8 +47,8 @@ export const getGoals = createAsyncThunk(
   "goals/getAll",
   async (_, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.getGoals(token);
+      const token = (thunkAPI.getState() as RootState)?.auth?.user?.token;
+      if (token) return await goalService.getGoals(token);
     } catch (error: any) {
       const message =
         (error.response &&
@@ -52,10 +65,10 @@ export const getGoals = createAsyncThunk(
 // Delete goal
 export const deleteGoal = createAsyncThunk(
   "goals/delete",
-  async (id, thunkAPI) => {
+  async (id: string, thunkAPI) => {
     try {
-      const token = thunkAPI.getState().auth.user.token;
-      return await goalService.deleteGoal(id, token);
+      const token = (thunkAPI.getState() as RootState)?.auth?.user?.token;
+      if (token) return await goalService.deleteGoal(id, token);
     } catch (error: any) {
       const message =
         (error.response &&
@@ -73,7 +86,7 @@ export const goalSlice = createSlice({
   name: "goal",
   initialState,
   reducers: {
-    reset: (state) => initialState,
+    reset: () => initialState,
   },
   extraReducers: (builder) => {
     builder
@@ -122,4 +135,5 @@ export const goalSlice = createSlice({
 });
 
 export const { reset } = goalSlice.actions;
+
 export default goalSlice.reducer;
